Updates to verbose level of dynamixel interface.  Support for iOS and Android being implemented.

+++ edisonLibmogiPackage/libmogi/simulation/src/Mesh.cpp
@@ -58,8 +58,7 @@ MBmesh::~MBmesh() {
 //	return *this;
 //}
 
-int MBmesh::drawWithMaterials(MBshader* shader,
-		std::vector<MBmaterial*>& materials) {
+int MBmesh::drawWithMaterials(MBshader* shader, std::vector<MBmaterial*>& materials) {
 	materials[materialIndex]->sendToShader(shader);
 
 	return draw(shader);
@@ -110,7 +109,7 @@ int MBmesh::draw(MBshader* shader) {
 	glVertexAttribPointer(UV, 2, GL_FLOAT, GL_FALSE, sizeof(VertexData),
 			(void*) (12 * sizeof(float)));
 
-	glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+	glDrawElements(GL_TRIANGLES, (int)indices.size(), GL_UNSIGNED_INT, 0);
 
 	glDisableVertexAttribArray(vertex);
 	glDisableVertexAttribArray(normal);
@@ -145,7 +144,7 @@ int MBmesh::draw(MBshader* shader) {
 	}
 
 	int MBmesh::getNumberOfVertices() {
-		return data.size();
+		return (int)data.size();
 	}
 
 	void MBmesh::setColor(float red, float green, float blue) {
@@ -158,12 +157,10 @@ int MBmesh::draw(MBshader* shader) {
 
 	void MBmesh::loadVerticesToVertexBufferObject() {
 		glBindBuffer(GL_ARRAY_BUFFER, vertexBufferObject);
-		glBufferData(GL_ARRAY_BUFFER, data.size() * sizeof(VertexData), &data[0],
-					 GL_STATIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, data.size() * sizeof(VertexData), &data[0], GL_STATIC_DRAW);
 
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, indexBufferObject);
-		glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(unsigned int),
-					 &indices[0], GL_STATIC_DRAW);
+		glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(unsigned int), &indices[0], GL_STATIC_DRAW);
 
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
